requirements.txt
# 数据处理
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# 机器学习
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# 深度学习 (可选)
torch>=2.0.0
tensorflow>=2.13.0

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 金融数据处理
# ta-lib>=0.4.0
yfinance>=0.2.0
tushare>=1.2.0

# 特征工程
feature-engine>=1.6.0
category-encoders>=2.6.0

# 模型解释性
shap>=0.42.0
lime>=0.2.0

# 超参数优化
optuna>=3.2.0
hyperopt>=0.2.0

# 实用工具
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0

# 开发工具
jupyter>=1.0.0
notebook>=6.5.0
ipywidgets>=8.0.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0 